inline renderAgentMCPAction
